+++ src/index.ts
@@ -1,6 +1,6 @@
-import { Tree } from "./tree";
-import { OneToManyMap } from "./one-to-many-map";
-import { BiMap } from "./bi-map";
+import { Tree } from "./tree/tree";
+import { OneToManyMap } from "./one-to-many-map/one-to-many-map";
+import { BiMap } from "./bi-map/bi-map";
 export { BiMap };
 export { OneToManyMap };
 export { Tree };
